Move `asSyncAdapter` to `EventsContract` and `DmfsTasksContract`
