+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/calendar/AndroidCalendar.kt
@@ -21,6 +21,7 @@ import androidx.core.content.contentValuesOf
 import at.bitfire.synctools.storage.BatchOperation.CpoBuilder
 import at.bitfire.synctools.storage.LocalStorageException
 import at.bitfire.synctools.storage.UnknownProperty
+import at.bitfire.synctools.storage.calendar.EventsContract.asSyncAdapter
 import at.bitfire.synctools.storage.toContentValues
 import org.jetbrains.annotations.TestOnly
 import java.util.LinkedList
