+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/calendar/AndroidCalendarProvider.kt
@@ -18,6 +18,7 @@ import androidx.core.content.contentValuesOf
 import at.bitfire.synctools.icalendar.Css3Color
 import at.bitfire.synctools.storage.LocalStorageException
 import at.bitfire.synctools.storage.calendar.AndroidCalendarProvider.Companion.COLUMN_CALENDAR_SYNC_STATE
+import at.bitfire.synctools.storage.calendar.EventsContract.asSyncAdapter
 import at.bitfire.synctools.storage.toContentValues
 import java.util.LinkedList
 import java.util.logging.Level
