--- synctools/src/main/kotlin/at/bitfire/synctools/storage/calendar/CalendarUtils.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
- */
-
-package at.bitfire.synctools.storage.calendar
-
-import android.accounts.Account
-import android.net.Uri
-import android.provider.CalendarContract
-
-fun Uri.asSyncAdapter(account: Account): Uri = buildUpon()
-    .appendQueryParameter(CalendarContract.Calendars.ACCOUNT_NAME, account.name)
-    .appendQueryParameter(CalendarContract.Calendars.ACCOUNT_TYPE, account.type)
-    .appendQueryParameter(CalendarContract.CALLER_IS_SYNCADAPTER, "true")
-    .build()
