+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/tasks/DmfsTasksContract.kt
@@ -4,12 +4,21 @@
 
 package at.bitfire.synctools.storage.tasks
 
+import android.accounts.Account
+import android.net.Uri
 import org.dmfs.tasks.contract.TaskContract
 
 /**
  * How synctools uses tasks.org task sync columns and data rows.
  */
 object DmfsTasksContract {
+
+    fun Uri.asSyncAdapter(account: Account): Uri = buildUpon()
+        .appendQueryParameter(TaskContract.ACCOUNT_NAME, account.name)
+        .appendQueryParameter(TaskContract.ACCOUNT_TYPE, account.type)
+        .appendQueryParameter(TaskContract.CALLER_IS_SYNCADAPTER, "true")
+        .build()
+
 
     /**
      * Custom sync column to store the last known ETag of a task.
