+++ synctools/src/test/kotlin/at/bitfire/synctools/storage/tasks/DmfsTasksContractTest.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
+ */
+
+package at.bitfire.synctools.storage.tasks
+
+import android.accounts.Account
+import android.net.Uri
+import at.bitfire.synctools.storage.tasks.DmfsTasksContract.asSyncAdapter
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+
+@RunWith(RobolectricTestRunner::class)
+class DmfsTasksContractTest {
+
+    @Test
+    fun testUriHelper_asSyncAdapter() {
+        val account = Account("testName", "testType")
+        val baseUri = Uri.parse("test://example.com/")
+        assertEquals(
+            Uri.parse("$baseUri?account_name=testName&account_type=testType&caller_is_syncadapter=true"),
+            baseUri.asSyncAdapter(account)
+        )
+    }
+
+}
